pyproject.toml
[project]
name = "report-genius"
version = "0.2.0"
description = "AI-powered portable view template system for Kahua construction management"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Proprietary"}
authors = [
    {name = "Kahua, Inc"}
]
keywords = ["kahua", "construction", "templates", "portable-view", "docx"]

dependencies = [
    # Core
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    
    # API
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.25.0",
    "httpx>=0.25.0",
    "websockets>=12.0",
    
    # Document generation
    "python-docx>=1.0.0",
    "matplotlib>=3.8.0",
    "python-dateutil>=2.8.0",
    
    # AI/LLM
    "langchain>=0.1.0",
    "langchain-anthropic>=0.1.0",
    "langgraph>=0.0.20",
    
    # PDF processing (for template analysis)
    "pymupdf>=1.23.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
    "mypy>=1.6.0",
]

[project.scripts]
# CLI entry point (to be implemented)
# report-genius = "report_genius.cli:main"

[project.entry-points."report_genius.renderers"]
docx = "report_genius.rendering:DocxRenderer"
legacy = "report_genius.rendering:LegacyRenderer"

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
include = ["report_genius*"]

[tool.setuptools.package-data]
"report_genius" = ["*.json", "*.md"]

[tool.ruff]
line-length = 100
target-version = "py310"
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
]

[tool.ruff.isort]
known-first-party = ["report_genius"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
